Fixed the progressBar for midi files going to zero on pauses
The issue was that we were basing the progressBar off the currentTime variable and for midi files it was checking if tinySynth.getPlayTime existed, which it doesn't. I changed that to getPlayStatus and changed the synthTime to be calculated based off the ticks instead of doing more complex math to time how long the song had been paused for vs the time the audioPlayer up time. Just trust me, it works now.